Fix required properties.
According to the spec, `required` is an object-level attribute, not a property attribute.
Before this change, we were setting objects as `required` whenever they
had a list of `required` properties. e.g.:

```
CreateCallRequestBase:
  type: object
  required:
    - to
  properties:
    to:
      type: array
      items:
        oneOf:
          - "$ref": "#/components/schemas/EndpointPhoneTo"
          - "$ref": "#/components/schemas/EndpointSip"
          - "$ref": "#/components/schemas/EndpointWebsocket"
          - "$ref": "#/components/schemas/EndpointVBCExtension"
    from:
      "$ref": "#/components/schemas/EndpointPhoneFrom"

EndpointPhoneFrom:
  type: object
  description: Connect to a Phone (PSTN) number
  required:
    - number
    - type
  properties:
    type:
      type: string
      example: phone
      description: The type of connection. Must be `phone`
    number:
      "$ref": "#/components/schemas/AddressE164"
```

In the previous case, the property `from` was being marked as
`required`, even though it isn't listed under `required`.
This commit fixes this issue by checking at object-level.